Player and Book Interaction Modified the Left side of the Throne Created a Trigger Script to allow the pop up of the BackStory Imported a few PNG files to use for the UI Instructions documenting the process were included Folder titled 'Alissen' was created to keep track of my Jira tasks
