+++ Assets/Alissen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2d04dbb68881ff40b6ceb640913ba46
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
