+++ Assets/Alissen/Book_Interaction/Trigger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fef85142ac0c2424782dc9d2649d5b55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
