+++ Assets/Alissen/HealthBar Files/Billboard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 289b48e7dcc1d244b82945f99a4ea14f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
