+++ Assets/Alissen/REPORT_INSTRUCTIONS/README_HEALTHBAR.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class README_HEALTHBAR : MonoBehaviour
+{
+   /*
+    On the 3D Character, right Click and select: UI->Canvas
+    Rename to Healthbar Canvas
+
+
+
+
+   */
+}
