--- Assets/Book_Interaction.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cbe1c5a957c74c74ba2cecff40396bf1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
