--- Assets/HealthBar Files.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 57015a93af3dae44c9675e5d904dcd4c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
