--- Assets/HealthBar Files/Billboard.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Billboard : MonoBehaviour
-{
-    public Transform cam;
-    // Start is called before the first frame update
-    
-
-    // Update is called once per frame
-    void LateUpdate()
-    {
-        transform.LookAt(transform.position+cam.forward);
-    }
-}
