--- Assets/HealthBar Files/HealthBar.cs
@@ -1,22 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-public class HealthBar : MonoBehaviour
-{
-    public Slider slider;
-    public Gradient gradient;
-    public Image fill;
-    // Start is called before the first frame update
-    public void SetMaxHealth(int health)
-    {
-        slider.maxValue=health;
-        slider.value=health;
-        fill.color=gradient.Evaluate(slider.normalizedValue); // removed the 1f
-    }
-public void SetHealth(int health)
-{
-   slider.value=health; 
-   fill.color=gradient.Evaluate(slider.normalizedValue);
-}
-}
