--- Assets/HealthBar Files/HealthBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0eb2c511077a4e342be497abdc2fe14f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
