--- Assets/HealthBar Files/Heart.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:50faa2ae0775d9dd86803ed41257cdaf90dc73b34e9c636ef081eb1828d4c0b8
-size 2430
