--- Assets/HealthBar Files/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2f1647533ae194b41aefc97ef8dfe32d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
