--- Assets/REPORT_INSTRUCTIONS.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f53e8ab8fb1d08149a7f4d9699a46270
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
