--- Assets/REPORT_INSTRUCTIONS/README_HEALTHBAR.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dfcfcf3949573b14eb643aa03baad0a5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
