+++ ProjectSettings/AudioManager.asset
@@ -10,10 +10,6 @@ AudioManager:
   Default Speaker Mode: 2
   m_SampleRate: 0
   m_DSPBufferSize: 1024
-<<<<<<< HEAD
-  m_DSPBufferSize: 0
-=======
->>>>>>> Diego's_Prototype
   m_VirtualVoiceCount: 512
   m_RealVoiceCount: 32
   m_EnableOutputSuspension: 1
