+++ ProjectSettings/EditorSettings.asset
@@ -21,14 +21,6 @@ EditorSettings:
   m_ProjectGenerationRootNamespace: 
   m_EnableTextureStreamingInEditMode: 1
   m_EnableTextureStreamingInPlayMode: 1
-<<<<<<< HEAD
-  m_AsyncShaderCompilation: 1
-  m_CachingShaderPreprocessor: 1
-  m_PrefabModeAllowAutoSave: 1
-  m_EnterPlayModeOptionsEnabled: 0
-  m_EnterPlayModeOptions: 3
-=======
->>>>>>> Diego's_Prototype
   m_EnableEditorAsyncCPUTextureLoading: 0
   m_AsyncShaderCompilation: 1
   m_PrefabModeAllowAutoSave: 1
