+++ ProjectSettings/Physics2DSettings.asset
@@ -42,10 +42,6 @@ Physics2DSettings:
   m_QueriesHitTriggers: 1
   m_QueriesStartInColliders: 1
   m_CallbacksOnDisable: 1
-<<<<<<< HEAD
-  m_ReuseCollisionCallbacks: 0
-=======
->>>>>>> Diego's_Prototype
   m_ReuseCollisionCallbacks: 1
   m_AutoSyncTransforms: 0
   m_AlwaysShowColliders: 0
