+++ ProjectSettings/ProjectSettings.asset
@@ -4,11 +4,7 @@
 PlayerSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 24
-<<<<<<< HEAD
-  productGUID: 688a37b082dcd414181ff13519f72304
-=======
-  productGUID: a3d5ebd44f1e2b047a160ea464080028
->>>>>>> Diego's_Prototype
+  productGUID: 01f575dcba075534bab03b7cc3d92fda
   AndroidProfiler: 0
   AndroidFilterTouchesWhenObscured: 0
   AndroidEnableSustainedPerformanceMode: 0
@@ -60,10 +56,6 @@ PlayerSettings:
   iosShowActivityIndicatorOnLoading: -1
   androidShowActivityIndicatorOnLoading: -1
   iosUseCustomAppBackgroundBehavior: 0
-<<<<<<< HEAD
-  iosAllowHTTPDownload: 1
-=======
->>>>>>> Diego's_Prototype
   allowedAutorotateToPortrait: 1
   allowedAutorotateToPortraitUpsideDown: 1
   allowedAutorotateToLandscapeRight: 1
@@ -127,12 +119,6 @@ PlayerSettings:
   switchNVNShaderPoolsGranularity: 33554432
   switchNVNDefaultPoolsGranularity: 16777216
   switchNVNOtherPoolsGranularity: 16777216
-<<<<<<< HEAD
-  switchNVNMaxPublicTextureIDCount: 0
-  switchNVNMaxPublicSamplerIDCount: 0
-  switchMaxWorkerMultiple: 8
-=======
->>>>>>> Diego's_Prototype
   switchGpuScratchPoolGranularity: 2097152
   switchAllowGpuScratchShrinking: 0
   switchNVNMaxPublicTextureIDCount: 0
@@ -162,10 +148,6 @@ PlayerSettings:
     enable360StereoCapture: 0
   isWsaHolographicRemotingEnabled: 0
   enableFrameTimingStats: 0
-<<<<<<< HEAD
-  enableOpenGLProfilerGPURecorders: 1
-=======
->>>>>>> Diego's_Prototype
   useHDRDisplay: 0
   D3DHDRBitDepth: 0
   m_ColorGamuts: 00000000
@@ -447,10 +429,6 @@ PlayerSettings:
   switchNativeFsCacheSize: 32
   switchIsHoldTypeHorizontal: 0
   switchSupportedNpadCount: 8
-<<<<<<< HEAD
-  switchEnableTouchScreen: 1
-=======
->>>>>>> Diego's_Prototype
   switchSocketConfigEnabled: 0
   switchTcpInitialSendBufferSize: 32
   switchTcpInitialReceiveBufferSize: 64
@@ -462,11 +440,6 @@ PlayerSettings:
   switchSocketInitializeEnabled: 1
   switchNetworkInterfaceManagerInitializeEnabled: 1
   switchPlayerConnectionEnabled: 1
-<<<<<<< HEAD
-  switchUseNewStyleFilepaths: 0
-  switchUseLegacyFmodPriorities: 1
-=======
->>>>>>> Diego's_Prototype
   switchUseNewStyleFilepaths: 1
   switchUseLegacyFmodPriorities: 0
   switchUseMicroSleepForYield: 1
@@ -665,10 +638,6 @@ PlayerSettings:
     m_VersionName: 
   apiCompatibilityLevel: 6
   activeInputHandler: 0
-<<<<<<< HEAD
-  windowsGamepadBackendHint: 0
-=======
->>>>>>> Diego's_Prototype
   cloudProjectId: 
   framebufferDepthMemorylessMode: 0
   qualitySettingsNames: []
